Change default time unit from `milliseconds` to `seconds` (#523)
* use seconds in mlperfdriver duration

* use seconds in dart_ffi_run_benchmark

* use seconds in task config file

* use seconds for saving min duration

* use seconds for measured duration

* use seconds for saving loadgen duration

* use elapsed.inSeconds

* fix linter

+++ flutter/cpp/mlperf_driver.h
@@ -44,7 +44,7 @@ class MlperfDriver : public ::mlperf::SystemUnderTest {
 
   // Runs MLPerf tests.
   void RunMLPerfTest(const std::string& mode, int min_query_count,
-                     int min_duration, int single_stream_expected_latency_ns,
+                     double min_duration, int single_stream_expected_latency_ns,
                      const std::string& output_dir);
 
   // A human-readable string for loggin purposes.
